Actions, Update Gradle, Update Dependencies, J21 support.

+++ src/test/java/net/minecraftforge/srg2source/test/MinecraftTest.java
@@ -1,20 +1,6 @@
 /*
- * Srg2Source
- * Copyright (c) 2020.
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation version 2.1
- * of the License.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ * Copyright (c) Forge Development LLC
+ * SPDX-License-Identifier: LGPL-2.1-only
  */
 
 package net.minecraftforge.srg2source.test;
@@ -27,9 +13,7 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.io.PrintStream;
-import java.lang.reflect.Constructor;
 import java.net.URI;
-import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
@@ -50,19 +34,15 @@ import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 import java.util.zip.ZipOutputStream;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.objectweb.asm.Type;
 
-import cpw.mods.modlauncher.LaunchPluginHandler;
-import cpw.mods.modlauncher.TransformStore;
-import cpw.mods.modlauncher.TransformationServiceDecorator;
-import cpw.mods.modlauncher.TransformingClassLoader;
-import cpw.mods.modlauncher.api.ITransformationService;
 import net.minecraftforge.srg2source.api.RangeApplierBuilder;
 import net.minecraftforge.srg2source.api.RangeExtractorBuilder;
-import net.minecraftforge.srg2source.asm.TransformationService;
+import net.minecraftforge.srg2source.api.SourceVersion;
 import net.minecraftforge.srg2source.extract.RangeExtractor;
+import net.minecraftforge.srg2source.util.TransformingUtil;
 import net.minecraftforge.srg2source.util.Util;
 import net.minecraftforge.srg2source.util.io.ZipInputSupplier;
 import net.minecraftforge.srgutils.IMappingFile;
@@ -77,8 +57,8 @@ public class MinecraftTest {
     //Run the command: gradlew [MC_VERSION]:projectJoinedApplyPatches
     //It will make all necessary files for this test.
 
-    private static final Path MCP_ROOT = Paths.get("Z:/Projects/MCP/MCPConfig/");
-    private static final String MC_VERSION = "1.16.2";
+    private static final Path MCP_ROOT = Paths.get("Z:/Projects/MCPConfig/");
+    private static final String MC_VERSION = "1.20.5-rc1";
     private static final OutputStream NULL_OUTPUT = new OutputStream() { public void write(int b) throws IOException { } };
 
     /**
@@ -90,17 +70,7 @@ public class MinecraftTest {
         if (!Files.exists(MCP_ROOT))
             return;
 
-        TransformStore transformStore = new TransformStore();
-        Constructor<TransformationServiceDecorator> ctr = TransformationServiceDecorator.class.getDeclaredConstructor(ITransformationService.class);
-        ctr.setAccessible(true);
-        TransformationServiceDecorator sd = ctr.newInstance(new TransformationService());
-        sd.gatherTransformers(transformStore);
-        Path[] targetPaths = new Path[] {
-            getClassRoot("org/eclipse/jdt/core/dom/CompilationUnitResolver"),
-            getClassRoot(getClass().getName()),
-            getClassRoot(RangeExtractor.class.getName())
-        };
-        TransformingClassLoader tcl = new TransformingClassLoader(transformStore, new LaunchPluginHandler(), targetPaths);
+        var tcl = TransformingUtil.createTransformer();
 
         Class<?> cls = Class.forName(getClass().getName() + "$Redefined", true, tcl);
 
@@ -113,18 +83,6 @@ public class MinecraftTest {
         System.out.println("Apply   Time: " + apply);
     }
 
-    private Path getClassRoot(String cls) {
-        URL url = this.getClass().getResource("/" + cls.replace('.', '/') + ".class");
-        if (url == null)
-            return null;
-        String path = url.toString().substring(0, url.toString().length() - cls.length() - 6);
-        if ("jar".equals(url.getProtocol()) && path.endsWith("!/"))
-            path = path.substring(4, path.length() - 2);
-        if (path.startsWith("file:"))
-            path = path.substring(6);
-        return Paths.get(path);
-    }
-
     public static class Redefined {
         private static final Pattern SRG_FINDER = Pattern.compile("((?:func_|m_)[0-9]+_[a-zA-Z_]*?|(?:field_|f_)[0-9]+_[a-zA-Z_]*?|p_[\\w]+_\\d+_)\\b");
         private final String mcver;
@@ -135,19 +93,29 @@ public class MinecraftTest {
             this.root = Paths.get("build/test/mcptest/" + mcver);
             if (!Files.exists(root))
                 Files.createDirectories(root);
+
+            Assertions.assertTrue(RangeExtractor.hasBeenASMPatched(), "Range Extractor not Patched!");
         }
 
         private Path mcpBuild(String path) {
             return MCP_ROOT.resolve("build/versions/" + mcver + '/' + path);
         }
         private Path mcpData(String path) {
-            return MCP_ROOT.resolve("versions/release/" + mcver + '/' + path);
+            var ret = MCP_ROOT.resolve("versions/release/" + mcver + '/' + path);
+            if (!Files.exists(ret)) {
+                // Could be a pre-release
+                int idx = mcver.indexOf('.', 2);
+                if (idx == -1)
+                    throw new IllegalStateException("Missing MCP Data");
+                ret = MCP_ROOT.resolve("versions/pre/" + mcver.substring(0, idx) + '/' + mcver + '/' + path);
+            }
+            return ret;
         }
 
         public long extract(boolean forceOld) throws Exception {
             List<Path> libraries = gatherLibraries();
             Path src_srg = createSourceJar();
-            Assert.assertTrue("Missing SRG Source jar: " + src_srg, Files.exists(src_srg));
+            Assertions.assertTrue(Files.exists(src_srg), "Missing SRG Source jar: " + src_srg);
 
             Path src_mcp = getMappedSrc(src_srg);
 
@@ -156,6 +124,7 @@ public class MinecraftTest {
                     .input(ZipInputSupplier.create(src_mcp, StandardCharsets.UTF_8))
                     .batch(!forceOld)
                     .output(root.resolve("extract_" + (forceOld ? "legacy" : "batched") + ".txt"))
+                    .sourceCompatibility(SourceVersion.JAVA_21)
                     .logger(log);
 
                 libraries.forEach(l -> {
@@ -174,7 +143,7 @@ public class MinecraftTest {
 
         public long apply() throws Exception {
             Path src_srg = createSourceJar();
-            Assert.assertTrue("Missing SRG Source jar: " + src_srg, Files.exists(src_srg));
+            Assertions.assertTrue(Files.exists(src_srg), "Missing SRG Source jar: " + src_srg);
 
             Path src_mcp = getMappedSrc(src_srg);
             Path mcp_to_srg = getMcpToSrg();
@@ -199,7 +168,7 @@ public class MinecraftTest {
                     for (String path : clean.gatherAll(".java")) {
                         String c = new String(Util.readStream(clean.getInput(path)), StandardCharsets.UTF_8);
                         String o = new String(Util.readStream(clean.getInput(path)), StandardCharsets.UTF_8);
-                        Assert.assertEquals("Output Mismatch: " + path, c, o);
+                        Assertions.assertEquals(c, o, "Output Mismatch: " + path);
                     }
                 }
                 return duration;
@@ -216,14 +185,18 @@ public class MinecraftTest {
             env.put("create", "true");
             URI uri = URI.create("jar:file:/" + target.toAbsolutePath().toString().replace('\\', '/'));
             System.out.println("jar:file:/" + target.toAbsolutePath().toString().replace('\\', '/'));
-            Path src = mcpData("projects/joined/src/main/java");
-            try (FileSystem zipfs = FileSystems.newFileSystem(uri, env);
-                Stream<Path> stream = Files.walk(src)) {
-                for (Path p : stream.filter(f -> !Files.isDirectory(f)).collect(Collectors.toList())) {
-                    Path zfile = zipfs.getPath(src.relativize(p).toString());
-                    if (!Files.exists(zfile.getParent()))
-                        Files.createDirectories(zfile.getParent());
-                    Files.copy(p, zfile, StandardCopyOption.REPLACE_EXISTING);
+            for (var src : new Path[] { mcpData("projects/joined/src/main/java/"), mcpData("projects/shared/") }) {
+                if (!Files.exists(src))
+                    continue;
+
+                try (FileSystem zipfs = FileSystems.newFileSystem(uri, env);
+                    Stream<Path> stream = Files.walk(src)) {
+                    for (Path p : stream.filter(f -> !Files.isDirectory(f)).collect(Collectors.toList())) {
+                        Path zfile = zipfs.getPath(src.relativize(p).toString());
+                        if (!Files.exists(zfile.getParent()))
+                            Files.createDirectories(zfile.getParent());
+                        Files.copy(p, zfile, StandardCopyOption.REPLACE_EXISTING);
+                    }
                 }
             }
 
@@ -232,11 +205,11 @@ public class MinecraftTest {
 
         private List<Path> gatherLibraries() throws IOException {
             Path cfg = mcpBuild("joined.fernflower.libs.txt");
-            Assert.assertTrue("Could not find library list: " + cfg, Files.exists(cfg));
+            Assertions.assertTrue(Files.exists(cfg), "Could not find library list: " + cfg);
 
             List<Path> libs = Files.readAllLines(cfg).stream().filter(l -> l.startsWith("-e="))
                     .map(l -> Paths.get(l.substring(3))).collect(Collectors.toList());
-            libs.forEach(l -> Assert.assertTrue("Missing Library: " + l, Files.exists(l)));
+            libs.forEach(l -> Assertions.assertTrue(Files.exists(l), "Missing Library: " + l));
             return libs;
         }
 
@@ -244,12 +217,11 @@ public class MinecraftTest {
         //For the time being we need to figure out how to NOT remap lambda arguments so the test will pass..
         private String makeName(String srg) {
             String[] pts = srg.split("_");
-            if ("func".equals(pts[0]))
-                return "m_" + pts[1] + '_';
-            else if ("field".equals(pts[0]))
-                return "f_" + pts[1] + '_';
-            else if ("p".equals(pts[0]))
-                return "a_" + pts[1] + '_' + pts[2] + '_';
+            switch (pts[0]) {
+                case "func", "m": return "a_" + pts[1] + '_';
+                case "field", "f": return "b_" + pts[1] + '_';
+                case "p": return "c_" + pts[1] + '_' + pts[2] + '_';
+            }
             throw new IllegalArgumentException("Unknown SRG: " + srg);
         }
 
@@ -341,21 +313,25 @@ public class MinecraftTest {
             List<String> ret = new ArrayList<>();
 
             Path static_mtds = mcpBuild("data/static_methods.txt");
-            Set<String> statics = Files.lines(static_mtds, StandardCharsets.UTF_8).map(l -> l.split("#")[0]).filter(l -> !l.trim().isEmpty()).collect(Collectors.toSet());
+            if (Files.exists(static_mtds)) {
+                Set<String> statics = Files.lines(static_mtds, StandardCharsets.UTF_8).map(l -> l.split("#")[0]).filter(l -> !l.trim().isEmpty()).collect(Collectors.toSet());
 
-            try (InputStream input = Files.newInputStream(mcpData("joined.tsrg"))) {
-                IMappingFile srg = IMappingFile.load(input);
-                srg.getClasses().stream().flatMap(c -> c.getMethods().stream())
-                .filter(m -> m.getMapped().startsWith("func_") && !m.getDescriptor().contains("()")).forEach(m ->
-                    ret.add(m.getParent().getMapped() + '.' + Redefined.this.makeName(m.getMapped()) + m.getMappedDescriptor() + "=|" +
-                        buildArgs(m.getMapped(), m.getDescriptor(), statics.contains(m.getMapped())))
-                );
+                try (InputStream input = Files.newInputStream(mcpData("joined.tsrg"))) {
+                    IMappingFile srg = IMappingFile.load(input);
+                    srg.getClasses().stream().flatMap(c -> c.getMethods().stream())
+                    .filter(m -> m.getMapped().startsWith("func_") && !m.getDescriptor().contains("()")).forEach(m ->
+                        ret.add(m.getParent().getMapped() + '.' + Redefined.this.makeName(m.getMapped()) + m.getMappedDescriptor() + "=|" +
+                            buildArgs(m.getMapped(), m.getDescriptor(), statics.contains(m.getMapped())))
+                    );
+                }
             }
 
             Path ctrs = mcpData("constructors.txt");
-            Files.lines(ctrs, StandardCharsets.UTF_8).map(l -> l.split(" ")).forEach(pts ->
-                ret.add(pts[1] + ".<init>" + pts[2] + "=|" + buildArgs(pts[0], pts[2], false))
-            );
+            if (Files.exists(ctrs)) {
+                Files.lines(ctrs, StandardCharsets.UTF_8).map(l -> l.split(" ")).forEach(pts ->
+                    ret.add(pts[1] + ".<init>" + pts[2] + "=|" + buildArgs(pts[0], pts[2], false))
+                );
+            }
 
             try (PrintStream out = new PrintStream(new FileOutputStream(target.toFile()))) {
                 ret.forEach(l -> out.print(l + "\n"));
